Have a clearer interface between BW and rest of game DLL
Though SNetGetPlayerNames is still being directly called and bw_1161
module does have some dependencies on rest of the code.
