+++ game/src/lib.rs
@@ -8,31 +8,28 @@ extern crate whack;
 mod app_messages;
 mod app_socket;
 mod bw;
+mod bw_1161;
 mod cancel_token;
 mod chat;
 mod forge;
 mod game_state;
 mod game_thread;
 mod network_manager;
-mod observing;
 mod rally_point;
 mod snp;
 mod storm;
 mod udp;
 mod windows;
 
-use std::ffi::{CStr, OsStr};
 use std::fs::File;
 use std::io;
 use std::path::PathBuf;
-use std::ptr::null_mut;
 use std::sync::Mutex;
 use std::time::{Duration, Instant};
 
 use lazy_static::lazy_static;
 use libc::c_void;
 use winapi::um::processthreadsapi::{GetCurrentProcess, TerminateProcess};
-use winapi::um::winnt::HANDLE;
 
 use crate::game_state::GameStateMessage;
 use crate::game_thread::GameThreadMessage;
@@ -185,11 +182,13 @@ pub extern "C" fn OnInject() {
 
     info!("Logging started");
     std::panic::set_hook(Box::new(panic_hook));
+    bw::set_bw_impl(Box::new(bw_1161::Bw1161));
     unsafe {
-        patch_game();
+        bw::with_bw(|bw| bw.patch_game());
     }
 }
 
+/// 1.16.1 calls LoadLibrary + GetProcAddress(SnpBind) on this dll to get networking functions.
 #[no_mangle]
 #[allow(non_snake_case)]
 pub unsafe extern "stdcall" fn SnpBind(index: u32, functions: *mut *const bw::SnpFunctions) -> u32 {
@@ -203,291 +202,6 @@ pub unsafe extern "stdcall" fn SnpBind(index: u32, functions: *mut *const bw::Sn
 
 lazy_static! {
     static ref PATCHER: Mutex<whack::Patcher> = Mutex::new(whack::Patcher::new());
-}
-
-unsafe fn patch_game() {
-    use observing::with_replay_flag_if_obs;
-
-    whack_export!(pub extern "system" CreateEventA(*mut c_void, u32, u32, *const i8) -> *mut c_void);
-    whack_export!(pub extern "system" DeleteFileA(*const i8) -> u32);
-    whack_export!(pub extern "system"
-        CreateFileA(*const i8, u32, u32, *mut c_void, u32, u32, *mut c_void) -> HANDLE
-    );
-
-    let mut active_patcher = PATCHER.lock().unwrap();
-    forge::init_hooks(&mut active_patcher);
-    snp::init_hooks(&mut active_patcher);
-
-    let mut exe = active_patcher.patch_exe(0x0040_0000);
-    bw::init_funcs(&mut exe);
-    bw::init_vars(&mut exe);
-    exe.hook_opt(bw::WinMain, entry_point_hook);
-    exe.hook(bw::GameInit, process_init_hook);
-    exe.hook_opt(bw::OnSNetPlayerJoined, game_thread::player_joined);
-    exe.hook_opt(bw::ChatCommand, chat::chat_command_hook);
-    exe.hook_opt(bw::ScrollScreen, scroll_screen);
-    // Rendering during InitSprites is useless and wastes a bunch of time, so we no-op it
-    exe.replace(bw::INIT_SPRITES_RENDER_ONE, &[0x90, 0x90, 0x90, 0x90, 0x90]);
-    exe.replace(bw::INIT_SPRITES_RENDER_TWO, &[0x90, 0x90, 0x90, 0x90, 0x90]);
-
-    exe.hook_closure(bw::MinimapCtrl_InitButton, |a, orig| {
-        with_replay_flag_if_obs(|| orig(a))
-    });
-    exe.hook_closure(bw::MinimapCtrl_ShowAllianceDialog, |orig| {
-        with_replay_flag_if_obs(|| orig())
-    });
-    exe.hook_closure(bw::DrawMinimap, |orig| {
-        with_replay_flag_if_obs(|| orig())
-    });
-    // Bw force refreshes the minimap every second?? why
-    // And of course the DrawMinimap call in it is inlined so it has to be hooked separately.
-    exe.hook_closure(bw::Minimap_TimerRefresh, |orig| {
-        with_replay_flag_if_obs(|| orig())
-    });
-    exe.hook_closure(bw::RedrawScreen, |orig| {
-        with_replay_flag_if_obs(|| orig())
-    });
-    exe.hook_closure(
-        bw::AllianceDialog_EventHandler,
-        |a, b, orig| with_replay_flag_if_obs(|| orig(a, b)),
-    );
-    exe.hook_closure(bw::GameScreenLeftClick, |a, orig| {
-        with_replay_flag_if_obs(|| orig(a))
-    });
-    exe.hook_closure(bw::PlaySoundAtPos, |a, b, c, d, orig| {
-        with_replay_flag_if_obs(|| orig(a, b, c, d))
-    });
-    exe.hook_closure(bw::DrawResourceCounts, |a, b, orig| {
-        with_replay_flag_if_obs(|| orig(a, b))
-    });
-    exe.hook_opt(bw::ProcessCommands, observing::process_commands_hook);
-    exe.hook_opt(bw::Command_Sync, observing::sync_command_hook);
-    exe.hook_opt(bw::ChatMessage, observing::chat_message_hook);
-    exe.hook_opt(bw::LoadDialog, observing::load_dialog_hook);
-    exe.hook_opt(bw::InitUiVariables, observing::init_ui_variables_hook);
-    exe.hook_opt(bw::UpdateCommandCard, observing::update_command_card_hook);
-    exe.hook_opt(
-        bw::CmdBtn_EventHandler,
-        observing::cmdbtn_event_handler_hook,
-    );
-    exe.hook_opt(bw::DrawCommandButton, observing::draw_command_button_hook);
-    exe.hook_opt(bw::GetGluAllString, observing::get_gluall_string_hook);
-    exe.hook_opt(
-        bw::UpdateNetTimeoutPlayers,
-        observing::update_net_timeout_players,
-    );
-    exe.hook_opt(
-        bw::CenterScreenOnOwnStartLocation,
-        observing::center_screen_on_start_location,
-    );
-
-    exe.import_hook_opt(&b"kernel32"[..], CreateEventA, create_event_hook);
-    exe.import_hook_opt(&b"kernel32"[..], DeleteFileA, delete_file_hook);
-    exe.import_hook_opt(&b"kernel32"[..], CreateFileA, create_file_hook);
-
-    // Check for a rare-but-dumb storm bug where the codegen for unrolled memcpy/blitting
-    // does an OOB string read and ends up generating broken code.
-    // (This data is initialized in storm's DllMain, so it has run already)
-    let surface_copy_code_ptr = *bw::storm::surface_copy_code;
-    if !surface_copy_code_ptr.is_null() {
-        let surface_copy_code = (*surface_copy_code_ptr).code_offsets[0xa0];
-        if *surface_copy_code.add(1) != 6 {
-            for i in 0..0xa0 {
-                *surface_copy_code.add(i * 0x10 + 0x1) = 0x6;
-                *surface_copy_code.add(i * 0x10 + 0x9) = 0x7;
-            }
-        }
-    }
-}
-
-fn scroll_screen(orig: unsafe extern fn()) {
-    if !forge::input_disabled() {
-        unsafe {
-            orig();
-        }
-    }
-}
-
-unsafe fn create_event_hook(
-    security: *mut c_void,
-    init_state: u32,
-    manual_reset: u32,
-    name: *const i8,
-    orig: unsafe extern fn(*mut c_void, u32, u32, *const i8) -> *mut c_void,
-) -> *mut c_void {
-    use winapi::um::errhandlingapi::SetLastError;
-    if !name.is_null() {
-        if CStr::from_ptr(name).to_str() == Ok("Starcraft Check For Other Instances") {
-            // BW just checks last error to be ERROR_ALREADY_EXISTS
-            SetLastError(0);
-            return null_mut();
-        }
-    }
-    orig(security, init_state, manual_reset, name)
-}
-
-fn ascii_path_filename(val: &[u8]) -> &[u8] {
-    val.rsplit(|&x| x == b'/' || x == b'\\')
-        .skip_while(|x| x.is_empty())
-        .next()
-        .unwrap_or_else(|| &[])
-}
-
-#[test]
-fn test_ascii_path_filename() {
-    assert_eq!(ascii_path_filename(b"asd/qwe/zxc"), b"zxc");
-    assert_eq!(ascii_path_filename(b"asd/qwe/z/c/"), b"c");
-    assert_eq!(ascii_path_filename(b"asd/qwe\\zxc.rep"), b"zxc.rep");
-    assert_eq!(ascii_path_filename(b"asd\\qwe//zxc"), b"zxc");
-    assert_eq!(ascii_path_filename(b"zxc"), b"zxc");
-    assert_eq!(ascii_path_filename(b"\\zxc"), b"zxc");
-    assert_eq!(ascii_path_filename(b"zxc///\\"), b"zxc");
-    assert_eq!(ascii_path_filename(b"\\zxc\\"), b"zxc");
-    assert_eq!(ascii_path_filename(b"\\/\\"), b"");
-}
-
-unsafe fn delete_file_hook(filename: *const i8, orig: unsafe extern fn(*const i8) -> u32) -> u32 {
-    if ascii_path_filename(CStr::from_ptr(filename).to_bytes()) == b"LastReplay.rep" {
-        // Before saving the last replay BW first tries to delete it, which can fail.
-        // We no-op it since we're saving the last replay ourselves.
-        1
-    } else {
-        orig(filename)
-    }
-}
-
-fn get_documents_path() -> Result<PathBuf, io::Error> {
-    use winapi::um::combaseapi::CoTaskMemFree;
-    use winapi::um::knownfolders::FOLDERID_Documents;
-    use winapi::um::shlobj::SHGetKnownFolderPath;
-
-    unsafe {
-        let mut path = null_mut();
-        let error = SHGetKnownFolderPath(&FOLDERID_Documents, 0, null_mut(), &mut path);
-        if error != 0 {
-            return Err(io::Error::from_raw_os_error(error));
-        }
-        let len = (0..).find(|&x| *path.add(x) == 0).unwrap();
-        let slice = std::slice::from_raw_parts(path, len);
-        let result = windows::os_string_from_winapi(slice);
-        CoTaskMemFree(path as *mut _);
-        Ok(result.into())
-    }
-}
-
-fn initial_number(path: &OsStr) -> u32 {
-    use std::os::windows::ffi::OsStrExt;
-    // Trickery to parse the initial number without assuming UTF-8 filename..
-    path.encode_wide()
-        .take_while(|&x| x >= b'0' as u16 && x <= b'9' as u16)
-        .fold(0, |old, new| {
-            old.wrapping_mul(10)
-                .wrapping_add((new - b'0' as u16) as u32)
-        })
-}
-
-#[test]
-fn test_initial_number() {
-    assert_eq!(initial_number(OsStr::new("123asdhk")), 123);
-    assert_eq!(initial_number(OsStr::new("asd")), 0);
-    assert_eq!(initial_number(OsStr::new("a234sd")), 0);
-    assert_eq!(initial_number(OsStr::new("1")), 1);
-    assert_eq!(initial_number(OsStr::new("001241.rep")), 1241);
-    assert_ne!(
-        initial_number(OsStr::new("100000000000000000001241.rep")),
-        0
-    );
-}
-
-unsafe fn create_file_hook(
-    filename: *const i8,
-    access: u32,
-    share_mode: u32,
-    security_attributes: *mut c_void,
-    creation_disposition: u32,
-    flags: u32,
-    template: *mut c_void,
-    orig: unsafe extern fn(*const i8, u32, u32, *mut c_void, u32, u32, *mut c_void) -> HANDLE,
-) -> HANDLE {
-    use winapi::um::fileapi::CreateFileW;
-    use winapi::um::handleapi::INVALID_HANDLE_VALUE;
-
-    if ascii_path_filename(CStr::from_ptr(filename).to_bytes()) != b"LastReplay.rep" {
-        return orig(
-            filename,
-            access,
-            share_mode,
-            security_attributes,
-            creation_disposition,
-            flags,
-            template,
-        );
-    }
-
-    let documents_path = match get_documents_path() {
-        Ok(o) => o,
-        Err(e) => {
-            error!("Couldn't retrieve user's document folder path: {}", e);
-            return INVALID_HANDLE_VALUE;
-        }
-    };
-    let replay_folder = documents_path.join("Starcraft\\maps\\replays\\Auto");
-    if !replay_folder.is_dir() {
-        if let Err(e) = std::fs::create_dir_all(&replay_folder) {
-            error!(
-                "Couldn't create replay folder '{}': {}",
-                replay_folder.display(),
-                e
-            );
-            return INVALID_HANDLE_VALUE;
-        }
-    }
-
-    let entries = match std::fs::read_dir(&replay_folder) {
-        Ok(o) => o,
-        Err(e) => {
-            error!(
-                "Couldn't read replay folder '{}': {}",
-                replay_folder.display(),
-                e
-            );
-            return INVALID_HANDLE_VALUE;
-        }
-    };
-    let mut count = 0;
-    for entry in entries {
-        let entry = match entry {
-            Ok(o) => o,
-            Err(e) => {
-                error!(
-                    "Couldn't read replay folder '{}': {}",
-                    replay_folder.display(),
-                    e
-                );
-                return INVALID_HANDLE_VALUE;
-            }
-        };
-        let path = entry.path();
-        if path.extension() == Some(OsStr::new("rep")) {
-            if let Some(file_stem) = path.file_stem() {
-                count = count.max(initial_number(file_stem));
-            }
-        }
-    }
-    let filename = format!(
-        "{:04}_{}.rep",
-        count + 1,
-        chrono::Local::now().format("%Y-%m-%d")
-    );
-    CreateFileW(
-        windows::winapi_str(replay_folder.join(filename)).as_ptr(),
-        access,
-        share_mode,
-        security_attributes as *mut _,
-        creation_disposition,
-        flags,
-        template,
-    )
 }
 
 unsafe fn entry_point_hook(
